moved JS incompatible RetainedInstanceStoreExt functions to jvm platform modules
