+++ appyx-navigation/common/src/androidMain/kotlin/com/bumble/appyx/navigation/store/RetainedInstanceStoreExt.kt
@@ -33,18 +33,3 @@ inline fun <reified T : Any> BuildContext.getRetainedInstance(
         disposer = disposer,
         factory = factory
     )
-
-/**
- * Obtains or creates an instance of a class using the identifier from the BuildContext.
- */
-inline fun <reified T : Any> BuildContext.getRetainedInstance(
-    key: String,
-    noinline disposer: (T) -> Unit = {},
-    noinline factory: () -> T
-) =
-    RetainedInstanceStore.get(
-        storeId = identifier,
-        key = key,
-        disposer = disposer,
-        factory = factory
-    )
