+++ appyx-navigation/common/src/desktopMain/kotlin/com/bumble/appyx/navigation/store/RetainedInstanceStoreExt.kt
@@ -1,0 +1,35 @@
+package com.bumble.appyx.navigation.store
+
+import com.bumble.appyx.navigation.modality.BuildContext
+
+/**
+ * Obtains or creates an instance of a class using the class name as the key.
+ * If you need multiple instances of an object with the same key, do not use this extension.
+ */
+inline fun <reified T : Any> RetainedInstanceStore.get(
+    storeId: String,
+    noinline disposer: (T) -> Unit = {},
+    noinline factory: () -> T
+): T =
+    get(
+        storeId = storeId,
+        key = T::class.java.name,
+        disposer = disposer,
+        factory = factory
+    )
+
+/**
+ * Obtains or creates an instance of a class using the class name as the key, and uses the
+ * identifier from the BuildContext.
+ *
+ * If you need multiple instances of an object with the same key, do not use this extension.
+ */
+inline fun <reified T : Any> BuildContext.getRetainedInstance(
+    noinline disposer: (T) -> Unit = {},
+    noinline factory: () -> T
+) =
+    RetainedInstanceStore.get(
+        storeId = identifier,
+        disposer = disposer,
+        factory = factory
+    )
